Code quality: Removed blank spaces, extracted ConfirmationConfiguration to a separate class
