+++ TestProjects/Android-Support/Fragments/Example.Core/Example.Core.csproj
@@ -49,6 +49,7 @@
     <Compile Include="ViewModels\ExampleRecyclerViewModel.cs" />
     <Compile Include="ViewModels\SecondHostViewModel.cs" />
     <Compile Include="ViewModels\ConfirmationViewModel.cs" />
+    <Compile Include="Model\ConfirmationConfiguration.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildExtensionsPath32)\Microsoft\Portable\$(TargetFrameworkVersion)\Microsoft.Portable.CSharp.targets" />
   <ItemGroup>
