+++ TestProjects/Android-Support/Fragments/Example.Core/Model/ConfirmationConfiguration.cs
@@ -1,0 +1,11 @@
+﻿using System;
+namespace Example.Core.Model
+{
+    public class ConfirmationConfiguration
+    {
+        public string Title { get; set; }
+        public string Body { get; set; }
+        public string PositiveCommandText { get; set; }
+        public string NegativeCommandText { get; set; }
+    }
+}
