+++ TestProjects/Android-Support/Fragments/Example.Droid/Fragments/ConfirmationFragment.cs
@@ -49,5 +49,4 @@ namespace Example.Droid.Fragments
             }
         }
     }
-
 }
